set reqwest to default-features = false (#12)
Hi there!

I'm in the process of trying to get OpenSSL out of my dependency
closure, in favor of `rustls`.

Currently `jwks` implicitly enables all default features of `reqwest`,
which implicitly enables `hyper-tls`, `openssl-sys`, etc.

It would be great if it could have the default features disabled so that
the caller can pick which backend to include.

This is technically a potentially breaking change, if both of the
following criteria apply:
- caller is using a HTTPS URI and the implicitly created reqwest client
- caller does NOT have reqwest in their direct dependencies, or they DO,
but have default features disabled and no tls backend enabled

I'd be glad to cut this as a feature flag instead, or as a version bump.
Open to suggestions!